авторизация

+++ src/main/java/com/pet/therapy/db/repo/DiaryRepo.java
@@ -3,5 +3,5 @@ package com.pet.therapy.db.repo;
 import com.pet.therapy.db.entity.DiaryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiaryRepo extends JpaRepository<DiaryEntity, Integer> {
+public interface DiaryRepo extends JpaRepository<DiaryEntity, Long> {
 }
